chore(collect): refactor state configuration, replace `AppStateDraft` with `AppState`, modularize configuration files, and improve the initialization flow
