+++ crates/collect/src/configure/atom.rs
