+++ crates/collect/src/lib.rs
@@ -1,8 +1,8 @@
 #![warn(clippy::all, clippy::pedantic, clippy::nursery)]
 
 pub mod cli;
+mod configure;
 mod services;
 mod source;
 mod state;
-mod state_config;
 mod utils;
